Move the extensions inside the same .cs
