--- Assets/Mirror/Runtime/NetworkReaderExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f9391710086644fceb3acd586899eae6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
