--- Assets/Mirror/Runtime/NetworkWriterExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: daea2d57ed16745589460a03a3ec10d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
